dvc: move GitTree to dvc/tree/
