+++ tests/func/test_tree.py
@@ -5,7 +5,7 @@ from dvc.path_info import PathInfo
 from dvc.repo import Repo
 from dvc.repo.tree import RepoTree
 from dvc.scm import SCM
-from dvc.scm.git import GitTree
+from dvc.tree.git import GitTree
 from dvc.tree.local import LocalRemoteTree
 from dvc.utils.fs import remove
 from tests.basic_env import TestDir, TestGit, TestGitSubmodule
